fix(wizard): probe accepts fractional scale_factor
ProbeScriptRequest.scale_factor was int32, so probing with a fractional
scale (e.g. TPCH smoke 0.01) failed server-side: 'invalid value for int32
field scaleFactor: 0.01'. Make it a double to match
domain.Workload.Parameters.scale_factor, and thread float64 through the
probe input → prober (SCALE_FACTOR env now formatted with FormatFloat
instead of %d).

+++ internal/services/test_wizard/service.go
@@ -104,7 +104,7 @@ type ProbeInput struct {
 	SQL          string
 	DriverType   string
 	PoolSize     int
-	ScaleFactor  int
+	ScaleFactor  float64
 	Env          map[string]string
 	Files        []ProbeFile
 	IncludeHuman bool
@@ -434,7 +434,7 @@ func (s *TestWizardService) ProbeScript(ctx context.Context, req *api.ProbeScrip
 		SQL:          req.GetSql(),
 		DriverType:   req.GetDriverType(),
 		PoolSize:     int(req.GetPoolSize()),
-		ScaleFactor:  int(req.GetScaleFactor()),
+		ScaleFactor:  req.GetScaleFactor(),
 		Env:          req.GetEnv(),
 		Files:        files,
 		IncludeHuman: req.GetIncludeHuman(),
